+++ CHANGELOG.md
@@ -4,8 +4,7 @@ x.x.x Release notes (yyyy-MM-dd)
 * None.
 
 ### Fixed
-* <How to hit and notice issue? what was the impact?> ([#????](https://github.com/realm/realm-js/issues/????), since v?.?.?)
-* None.
+* Fix TypeScript definition of `Realm.Dictionary.remove()`. (since v10.6.0)
 
 ### Compatibility
 * MongoDB Realm Cloud.